requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.26.2
sentence-transformers==2.2.2
rank-bm25==0.2.2
python-Levenshtein==0.23.0

# LangChain & AI
langchain==0.1.0
langchain-community==0.0.10
langchain-upstage==0.1.0
faiss-cpu==1.7.4

# Vector Database
pinecone-client==3.0.0

# Database
pymongo==4.6.0
redis==5.0.1

# Korean NLP
konlpy==0.6.0
nltk==3.8.1

# Utilities
pytz==2023.3
ipython==8.18.1
